Templatize RealSenseParamManager instead of inheritance
