+++ realsense2_camera/include/realsense2_camera/realsense_nodelet.h
@@ -1,31 +1,14 @@
 // License: Apache 2.0. See LICENSE file in root directory.
 // Copyright(c) 2018 Intel Corporation. All Rights Reserved
 
-#pragma once
+#ifndef REALSENSE2_CAMERA_REALSENSE_NODELET_H
+#define REALSENSE2_CAMERA_REALSENSE_NODELET_H
 
-#include <pluginlib/class_list_macros.h>
-#include <nodelet/nodelet.h>
-#include <image_transport/image_transport.h>
 #include <ros/ros.h>
-#include <ros/package.h>
-#include <librealsense2/rs.hpp>
-#include <librealsense2/rsutil.h>
-#include <librealsense2/hpp/rs_processing.hpp>
-#include <librealsense2/rs_advanced_mode.hpp>
-#include <cv_bridge/cv_bridge.h>
-#include <sensor_msgs/CameraInfo.h>
-#include <sensor_msgs/PointCloud2.h>
-#include <sensor_msgs/point_cloud2_iterator.h>
-#include <realsense2_camera/constants.h>
-#include <realsense2_camera/Extrinsics.h>
-#include <tf/transform_broadcaster.h>
-#include <tf2_ros/static_transform_broadcaster.h>
-#include <sensor_msgs/Imu.h>
-#include <realsense2_camera/IMUInfo.h>
+#include <nodelet/nodelet.h>
+#include <memory>
+
 #include <realsense2_camera/realsense_node.h>
-#include <csignal>
-#include <eigen3/Eigen/Geometry>
-#include <fstream>
 
 namespace realsense2_camera
 {
@@ -47,4 +30,6 @@ namespace realsense2_camera
         void tryGetLogSeverity(rs2_log_severity& severity) const;
         std::unique_ptr<RealSenseNode> _realSenseNode;
     };
-}//end namespace
+}  // namespace realsense2_camera
+
+#endif  // REALSENSE2_CAMERA_REALSENSE_NODE_H
