+++ realsense2_camera/package.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0"?>
 <package>
   <name>realsense2_camera</name>
-  <version>2.0.9</version>
+  <version>2.0.10</version>
   <description>RealSense Camera package allowing access to Intel 3D D400 cameras</description>
   <maintainer email="sergey.dorodnicov@intel.com">Sergey Dorodnicov</maintainer>
   <maintainer email="itay.carpis@intel.com">Itay Carpis</maintainer>
